requirements.txt
# experiment management
hydra-core
pytest
black==22.3.0
pre-commit>=3.7
mlflow-skinny # can optionally install extras, see https://pypi.org/project/mlflow-skinny/

# standard science stack
numpy
matplotlib

# deep learning
torch>=2.0.0
torch-geometric>=2.4.0
torch-ema
pytorch_optimizer>=3.2.0

# project-specific
torchdiffeq
xformers # might cause troubles, see https://github.com/facebookresearch/xformers
pelican-lite
lgatr[xformers_attention]>=1.3.2
lloca[xformers_attention]>=1.3.1

# weaver packages
tqdm>=4.41.0
PyYAML>=5.4
awkward0>=0.15.5
uproot>=4.2.0,<5.2.0
awkward>=1.8.0
lz4
xxhash
